normalize to "copyright" and "trademark" rather than single-char variants
This improves matching correctness when trying to match a license file
to the text of a license which includes "copyright" verbatim in the
body of the license text, since the same normalization is not applied
to the original license texts. This includes common licenses such as
the 2- and 3-clause BSD licenses.

Signed-off-by: Tal Einat <taleinat@gmail.com>